Fix build issues when compiling for a Linux host. (#45)
builder:
Fix file and directory permissions.
Add floating-point settings to assembler options.

cmd/csp-gen:
Fix file and directory permissions.

cmd/def-gen:
Fix file and directory permissions.

cmd/sigoc:
Fix file and directory permissions.

goir:
Replace usages of intptr_t with int in C API to better translate to Go when generating the CGo API using SWIG.

sigo runtime:
Fixed case errors in imports.

other:
Improvements to LLVM and MLIR SWIG interface files.
Removed X86 dialect libraries from link.rsp that are NOT available in the normal Linux build of MLIR by default.
Updated README.

+++ goir/include/Go-c/mlir/Dialects.h
@@ -22,7 +22,7 @@ extern "C"
 
   void mlirStringRefDestroy(MlirStringRef* ref);
 
-  intptr_t mlirTypeHash(MlirType type);
+  int mlirTypeHash(MlirType type);
 
   MlirAttribute mlirGoCreateTypeMetadata(MlirType type, MlirAttribute dict);
 
@@ -58,7 +58,7 @@ extern "C"
 
   MlirAttribute mlirGoCreateTypeMetadataDictionaryAttr(
     MlirContext context,
-    intptr_t nEntries,
+    int nEntries,
     MlirAttribute* entries);
 
   MlirBlock mlirRegionGetLastBlock(MlirRegion region);
@@ -73,11 +73,11 @@ extern "C"
   MlirOperation mlirValueGetDefiningOperation(MlirValue value);
 
   MlirBlock
-  mlirBlockCreate2(intptr_t nArgs, MlirType* args, intptr_t nLocations, MlirLocation* locations);
+  mlirBlockCreate2(int nArgs, MlirType* args, int nLocations, MlirLocation* locations);
 
   MlirAttribute mlirDistinctAttrGet(MlirAttribute attr);
 
-  void mlirGoBlockDumpTail(MlirBlock block, intptr_t count);
+  void mlirGoBlockDumpTail(MlirBlock block, int count);
 
   enum MlirGoAsmConstraintDirection
   {
